XR: Fix issue with effects and multiview
This amends 64821f3552aa7f3826de945141f1e3c16f175bc1 as the logic for
adjusting the view count in sub-views was done after the effects was
processed causing the queried view count to be out of sync.
This change also touches the general state tracking the layer node, to
make it more consistent and ensure the dependencies on the layer state
is processed in order.

Fixes: QTBUG-132811
Pick-to: 6.9 6.8
Change-Id: Idc4c78184f4af3834505c5d6c130e8212ee5837d
Reviewed-by: Andy Nichols <andy.nichols@qt.io>