Fix reentrancy issue in the renderer
In the case of nested View3Ds we're going to be calling into the
renderer again. That means, among other things, that the current
layer data will be changed and reset while the nested views are
rendered. To avoid this from blowing up we'll now store the state
of the renderer before we potentially start rendering sub-views
and then restore it again once we're done.

Pick-to: 6.9 6.8
Change-Id: Ie02d8ce9476858d17ab00b5fd14609a69479bb7b
Reviewed-by: Andy Nichols <andy.nichols@qt.io>