Fix uninitialized member in QQuick3DSceneRenderer
Caught by our internal code checker due to recent changes to this class
(unrelated). The member in question should always be updated as part of
the sync phase, so no known issue caused by it not being initialized in
the ctor. To avoid future problem the value now default to the enum
value "Unspecified".

Pick-to: 6.9 6.8
Change-Id: Ied2cf310155111aa91b209543708a25998203190
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>