Update the xr_interaction example
Add a 2D UI element to the example. Add XrItemHandle as a reusable
component, making it possible to move the XrItem around. Also improve
the appearance of XrCursor, and fix bug where cursor would not move
when doing a mouse drag.

Includes code cleanups and API improvements for the xr_shared
components.

Pick-to: 6.9
Change-Id: I6dd9e821f6d91e60704f56c5afc6f0e969d5d843
Reviewed-by: Christian Strømme <christian.stromme@qt.io>