Make 3D models block XrVirtualMouse event delivery
XrVirtualMouse works by doing a rayPickAll and iterate through
the results until it hits a QQuickItem that accepts the mouse
event. Before this patch, it would ignore any 3D Model encountered.
This would mean that interacting with 3D objects in front of a 2D
user interface could cause spurious mouse events.

The proper fix for this issue would be to integrate 3D objects in the
Qt Quick event delivery system, but that is a massive task. For now,
assume that if a Model has pickable set to true, then it will accept
interaction events, and therefore should block mouse event delivery to
anything behind it.

Pick-to: 6.9
Change-Id: I8115b4ede79005a63ede1af6a7a32ee835e253a2
Reviewed-by: Christian Strømme <christian.stromme@qt.io>
Reviewed-by: Nicholas Bennett <nicholas.bennett@qt.io>