Directional Lights should always reset qt_lightAttenuation
Previously it was possible that if PointLights or SpotLights appear in a
scene before a DirectionalLight, when calculating the specular
contribution of a DirectionalLight the attentuation value from whatever
previous positional light would be used instead of the intended value of
1.0. DirectionalLights should not have attenuation values, so make sure
to always reset qt_lightAttenuation to 1.0 before calculating the
specular term.

It is possible that this will change behavior in existing applications,
so expect a few Lancelot regressions. This should be the new baseline.

Pick-to: 6.9
Change-Id: Ic6d099967d83e1aa20063c11a4c7b2c96b64c22a
Reviewed-by: Christian Strømme <christian.stromme@qt.io>